It's working! It's Working!!!

+++ server/controllers/BlogController.js
@@ -29,7 +29,8 @@ export default class BlogController {
 
     async getAll(req, res, next) {
         try {
-            let data = await _blogService.find({}).populate('authorId', 'name')
+            let data = await _blogService.find({})
+                .populate('author', 'name')
             return res.send(data)
         } catch (error) { next(error) }
 
@@ -38,7 +39,7 @@ export default class BlogController {
     async getById(req, res, next) {
         try {
             let data = await _blogService.findById(req.params.id)
-
+                .populate('author', 'name')
             if (!data) {
                 throw new Error("Invalid Id")
             }
@@ -48,8 +49,8 @@ export default class BlogController {
 
     async getComments(req, res, next) {
         try {
-            let data = await _commentService.find({ blog: req.params.id }).populate("blog", "name")
-
+            let data = await _commentService.find({ blogId: req.params.id })
+                .populate('author', 'name')
             return res.send(data)
         } catch (error) { next(error) }
     }
@@ -58,7 +59,7 @@ export default class BlogController {
     async create(req, res, next) {
         try {
             //NOTE the user id is accessable through req.body.uid, never trust the client to provide you this information
-            req.body.authorId = req.session.uid
+            req.body.author = req.session.uid
             let data = await _blogService.create(req.body)
             res.send(data)
         } catch (error) { next(error) }
@@ -66,7 +67,7 @@ export default class BlogController {
 
     async edit(req, res, next) {
         try {
-            let data = await _blogService.findOneAndUpdate({ _id: req.params.id, }, req.body, { new: true })
+            let data = await _blogService.findOneAndUpdate({ _id: req.params.id, author: req.session.uid }, req.body, { new: true })
             if (data) {
                 return res.send(data)
             }
@@ -78,7 +79,10 @@ export default class BlogController {
 
     async delete(req, res, next) {
         try {
-            await _blogService.findOneAndRemove({ _id: req.params.id })
+            let data = await _blogService.findOneAndRemove({ _id: req.params.id, author: req.session.uid })
+            if (!data) {
+                throw new Error("invalid id")
+            }
             res.send("deleted blog")
         } catch (error) { next(error) }
